Move alertType and Alert module to lib
