--- frontend/src/metabase-lib/v1/Alert/index.ts
@@ -1,5 +1,0 @@
-export {
-  ALERT_TYPE_ROWS,
-  ALERT_TYPE_TIMESERIES_GOAL,
-  ALERT_TYPE_PROGRESS_BAR_GOAL,
-} from "./constants";
